PERA-3127 :: Use double instead of float for slippage tolerance (#467)

+++ app/src/main/kotlin/com/algorand/android/ui/swap/configuration/model/SwapConfigurationResult.kt
@@ -13,7 +13,7 @@
 package com.algorand.android.ui.swap.configuration.model
 
 data class SwapConfigurationResult(
-    val balancePercentage: Float?,
-    val slippageTolerance: Float?,
+    val balancePercentage: Int?,
+    val slippageTolerance: Double?,
     val useLocalCurrency: Boolean
 )
